CDI WELD SE, JPA HIBERNATE MYSQL, JSF MOJARRA PRIMEFACES

+++ src/main/java/com/github/pomona/domain/model/SubstanciaRepo.java
@@ -5,5 +5,5 @@ import com.github.common.domain.model.GenericoRepo;
 public interface SubstanciaRepo extends GenericoRepo<Substancia, SubstanciaId> {
     public Substancia substanciaPeloNome(String nome);
     
-    public int proximaOrdem();
+    public Long proximaOrdem();
 }
